Address Heather's review: extract renderSavedTrips and add XSS regression test
Heather's comment on PR #154 noted that the existing XSS regression coverage in tests/checklistRenderer.test.js only exercises checklist item / category rendering, not the saved-trip display path. The trip-name and destinationType render path lives in src/main.js and previously could not be unit-tested without running main.js's init() side effects.

- Extracted renderSavedTrips into src/savedTripsRenderer.js (the inline XSS safety comment travels with it).

- src/main.js now imports the function instead of defining it inline.

- Added tests/savedTripsRenderer.test.js with two jsdom regression cases — script/onerror payload as trip name, and as destinationType. Each asserts no script/img elements rendered, payload appears verbatim as text, and the side-effect a real XSS would produce did not happen.

- Generalized the eslint browser-globals override from the hard-coded checklistRenderer.test.js path to a tests/*Renderer.test.js glob so future DOM-rendering tests inherit it automatically.
